Rename Plugin/ to TcPlugin/
In preparation for introducin SrcPlugin/
