+++ large-anon/src/Data/Record/Anonymous/TcPlugin/Constraints/KnownFields.hs
@@ -3,7 +3,7 @@
 {-# LANGUAGE NamedFieldPuns  #-}
 {-# LANGUAGE RecordWildCards #-}
 
-module Data.Record.Anonymous.Plugin.Constraints.KnownFields (
+module Data.Record.Anonymous.TcPlugin.Constraints.KnownFields (
     CKnownFields(..)
   , parseKnownFields
   , solveKnownFields
@@ -15,10 +15,10 @@ import Data.Void
 import Data.Record.Anonymous.Internal.Row.KnownField (KnownField(..))
 import Data.Record.Anonymous.Internal.Row.KnownRow (KnownRow)
 import Data.Record.Anonymous.Internal.Row.ParsedRow (Fields)
-import Data.Record.Anonymous.Plugin.GhcTcPluginAPI
-import Data.Record.Anonymous.Plugin.NameResolution
-import Data.Record.Anonymous.Plugin.Parsing
-import Data.Record.Anonymous.Plugin.TyConSubst
+import Data.Record.Anonymous.TcPlugin.GhcTcPluginAPI
+import Data.Record.Anonymous.TcPlugin.NameResolution
+import Data.Record.Anonymous.TcPlugin.Parsing
+import Data.Record.Anonymous.TcPlugin.TyConSubst
 
 import qualified Data.Record.Anonymous.Internal.Row.FieldName as FieldName
 import qualified Data.Record.Anonymous.Internal.Row.KnownRow  as KnownRow
