+++ large-anon/src/Data/Record/Anonymous/TcPlugin/Constraints/Project.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE DataKinds       #-}
 {-# LANGUAGE RecordWildCards #-}
 
-module Data.Record.Anonymous.Plugin.Constraints.Project (
+module Data.Record.Anonymous.TcPlugin.Constraints.Project (
     CProject(..)
   , parseProject
   , solveProject
@@ -11,10 +11,10 @@ import Control.Monad (forM)
 import Data.Void
 
 import Data.Record.Anonymous.Internal.Row.ParsedRow (Fields)
-import Data.Record.Anonymous.Plugin.GhcTcPluginAPI
-import Data.Record.Anonymous.Plugin.NameResolution
-import Data.Record.Anonymous.Plugin.Parsing
-import Data.Record.Anonymous.Plugin.TyConSubst
+import Data.Record.Anonymous.TcPlugin.GhcTcPluginAPI
+import Data.Record.Anonymous.TcPlugin.NameResolution
+import Data.Record.Anonymous.TcPlugin.Parsing
+import Data.Record.Anonymous.TcPlugin.TyConSubst
 
 import qualified Data.Record.Anonymous.Internal.Row.KnownRow  as KnownRow
 import qualified Data.Record.Anonymous.Internal.Row.ParsedRow as ParsedRow
